rsa: Speed up key generation
Instead of randomly searching for a matching q for the chosen p,
compute a valid range of values for q such that pq has the requested bit
length. Choose a securely random value inside that range, and find a
nearby prime. This is the candidate for q. Binary search the space using
the candidate as a starting point until the product has the right number
of bits.